Add GUI for Rock Paper Scissors game
Add GUI for rock paper scissors game. Add buttons to make a player
choice. Add displays for scores, choices and rusults of rounds and the
game after one player reaches 5 points. Add freeze to game after game is
over and show a play again button which resets the score and un-freezes
the game.